[wip][fix] replace u32 to u64 in rv64/

+++ src/cpu/rv64/csr.rs
@@ -3,12 +3,12 @@ mod breakpoint;
 use super::{CSRname, PrivilegedLevel, TrapCause, Xstatus};
 use breakpoint::Triggers;
 
-const UMASK: u32 = 0b10000000000011010111100100110011;
-const SMASK: u32 = 0b10000000000011010111100100110011;
-const MMASK: u32 = 0b10000000011111111111100110111011;
+const UMASK: u64 = 0b10000000000011010111100100110011;
+const SMASK: u64 = 0b10000000000011010111100100110011;
+const MMASK: u64 = 0b10000000011111111111100110111011;
 
 pub struct CSRs {
-    csrs: [u32; 4096],
+    csrs: [u64; 4096],
     triggers: Triggers,
 }
 
@@ -30,7 +30,7 @@ impl CSRs {
         self
     }
 
-    pub fn bitset(&mut self, dist: Option<usize>, src: u32) {
+    pub fn bitset(&mut self, dist: Option<usize>, src: u64) {
         let mask = src;
         if mask != 0 {
             match dist.unwrap() {
@@ -41,7 +41,7 @@ impl CSRs {
         }
     }
 
-    pub fn bitclr(&mut self, dist: Option<usize>, src: u32) {
+    pub fn bitclr(&mut self, dist: Option<usize>, src: u64) {
         let mask = src;
         if mask != 0 {
             match dist.unwrap() {
@@ -52,7 +52,7 @@ impl CSRs {
         }
     }
 
-    pub fn write(&mut self, dist: Option<usize>, src: u32) {
+    pub fn write(&mut self, dist: Option<usize>, src: u64) {
         match dist.unwrap() {
             0x000 => self.csrs[0x300] = src & UMASK,
             0x100 => self.csrs[0x300] = src & SMASK,
@@ -61,7 +61,7 @@ impl CSRs {
         self.update_triggers(dist.unwrap(), src);
     }
 
-    fn read_xepc(&self, dist: usize) -> Result<u32, (Option<u32>, TrapCause, String)> {
+    fn read_xepc(&self, dist: usize) -> Result<u64, (Option<u64>, TrapCause, String)> {
         if self.csrs[CSRname::misa as usize] >> 2 & 0x1 == 1 {
             // C extension enabled (IALIGN = 16)
             Ok(self.csrs[dist] & !0b01)
@@ -71,7 +71,7 @@ impl CSRs {
         }
     }
 
-    pub fn read(&self, src: Option<usize>) -> Result<u32, (Option<u32>, TrapCause, String)> {
+    pub fn read(&self, src: Option<usize>) -> Result<u64, (Option<u64>, TrapCause, String)> {
         let dist = src.unwrap();
         match dist {
             0x000 => Ok(self.csrs[0x300] & UMASK),
@@ -81,9 +81,9 @@ impl CSRs {
         }
     }
 
-    pub fn read_xstatus(&self, priv_lv: PrivilegedLevel, xfield: Xstatus) -> u32 {
+    pub fn read_xstatus(&self, priv_lv: PrivilegedLevel, xfield: Xstatus) -> u64 {
         let xstatus = CSRname::mstatus as usize;
-        let mask: u32 = match priv_lv {
+        let mask: u64 = match priv_lv {
             PrivilegedLevel::Machine => MMASK,
             PrivilegedLevel::Supervisor => SMASK,
             PrivilegedLevel::User => UMASK,
@@ -111,9 +111,9 @@ impl CSRs {
         }
     }
 
-    pub fn write_xstatus(&mut self, priv_lv: PrivilegedLevel, xfield: Xstatus, data: u32) {
+    pub fn write_xstatus(&mut self, priv_lv: PrivilegedLevel, xfield: Xstatus, data: u64) {
         let xstatus = CSRname::mstatus as usize;
-        let mask: u32 = match priv_lv {
+        let mask: u64 = match priv_lv {
             PrivilegedLevel::Machine => MMASK,
             PrivilegedLevel::Supervisor => SMASK,
             PrivilegedLevel::User => UMASK,
