+++ src/cpu/rv64/decode/inst_32/base_i.rs
@@ -1,8 +1,8 @@
 use crate::cpu::instruction::OpecodeKind;
-use crate::cpu::rv32::decode::DecodeUtil;
+use crate::cpu::rv64::decode::DecodeUtil;
 use crate::cpu::TrapCause;
 
-pub fn parse_opecode(inst: u32) -> Result<OpecodeKind, &'static str> {
+pub fn parse_opecode(inst: u64) -> Result<OpecodeKind, &'static str> {
     let opmap: u8 = inst.slice(6, 0) as u8;
     let funct3: u8 = inst.slice(14, 12) as u8;
     let funct5: u8 = inst.slice(24, 20) as u8;
@@ -93,9 +93,9 @@ pub fn parse_opecode(inst: u32) -> Result<OpecodeKind, &'static str> {
 }
 
 pub fn parse_rd(
-    inst: u32,
+    inst: u64,
     opkind: &OpecodeKind,
-) -> Result<Option<usize>, (Option<u32>, TrapCause, String)> {
+) -> Result<Option<usize>, (Option<u64>, TrapCause, String)> {
     let rd: usize = inst.slice(11, 7) as usize;
 
     // B(EQ|NE|LT|GE|LTU|GEU), S(B|H|W), ECALL, EBREAK
@@ -133,9 +133,9 @@ pub fn parse_rd(
 }
 
 pub fn parse_rs1(
-    inst: u32,
+    inst: u64,
     opkind: &OpecodeKind,
-) -> Result<Option<usize>, (Option<u32>, TrapCause, String)> {
+) -> Result<Option<usize>, (Option<u64>, TrapCause, String)> {
     let rs1: usize = inst.slice(19, 15) as usize;
 
     // LUI, AUIPC, JAL, FENCE, ECALL, EBREAK
@@ -179,9 +179,9 @@ pub fn parse_rs1(
 }
 
 pub fn parse_rs2(
-    inst: u32,
+    inst: u64,
     opkind: &OpecodeKind,
-) -> Result<Option<usize>, (Option<u32>, TrapCause, String)> {
+) -> Result<Option<usize>, (Option<u64>, TrapCause, String)> {
     let rs2: usize = inst.slice(24, 20) as usize;
 
     // LUI, AUIPC, JAL, JALR L(B|H|W|BU|HU),
@@ -213,9 +213,9 @@ pub fn parse_rs2(
 
 #[allow(non_snake_case)]
 pub fn parse_imm(
-    inst: u32,
+    inst: u64,
     opkind: &OpecodeKind,
-) -> Result<Option<i32>, (Option<u32>, TrapCause, String)> {
+) -> Result<Option<i32>, (Option<u64>, TrapCause, String)> {
     let U_type = || (inst.slice(31, 12) << 12) as i32;
     let I_type = || {
         let imm32 = inst.slice(31, 20) as i32;
